refactor: refactor the home section
