+++ src/app/page.tsx
@@ -1,6 +1,6 @@
 import { AboutSection } from "@/components/sections/about-section";
 import { ContactSection } from "@/components/sections/contact-section";
-import { HomeSection } from "@/components/sections/home-section";
+import { HomeSection } from "@/components/sections/home/home-section";
 import { ProjectsSection } from "@/components/sections/projects-section";
 
 export default function Home() {
